requirements.txt

fastapi
uvicorn
streamlit

langchain
langchain-openai
langchain-community


chromadb


PyMuPDF  
beautifulsoup4  


google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client


python-dotenv
pydantic
requests
